Move origin.js to top level
Reasons for doing this:

1. Improving developer experience by allowing them to run `npm install`
from the repo root. We could hack our way around this with the old
scheme, but that gets weird really fast

2. Maintaining and exporting just a single npm package. What will be
consumed is a simple npm package. That's what should be at the top
level.

3. Simpler! This setup is more readable imo.
