+++ .gitignore
@@ -1,2 +1,39 @@
-.idea
+# See http://help.github.com/ignore-files/ for more about ignoring files.
+
+# Truffle build artifacts
+# https://ethereum.stackexchange.com/questions/19486/storing-a-truffle-contract-interface-in-version-control
+/build
+
+# compiled output
+/dist
+
+# dependencies
+/node_modules
+
+# IDEs and editors
+.idea/
+/.idea
+/.idea
+.project
+.classpath
+.c9/
+*.launch
+.settings/
+*.sublime-workspace
+
+# IDE - VSCode
+.vscode/*
+!.vscode/settings.json
+!.vscode/tasks.json
+!.vscode/launch.json
+!.vscode/extensions.json
+
+# misc
+/.sass-cache
+/connect.lock
+/coverage
+/libpeerconnection.log
+npm-debug.log
+testem.log
+/typings
 .DS_Store
