+++ contracts/test/TestUserRegistry.js
