+++ src/ResourceBase.js
