+++ src/resources/listings.js
