+++ test/contract-service.test.js
