+++ test/resource_listings.test.js
