+++ test/resource_purchases.test.js
