+++ test/user-registry-service.test.js
